rename file "model" file to mobile_builds
